refactor common folder in tests
